--- userCreate.js
@@ -1,2 +1,0 @@
-// const userCreateForm = document.querySelector('#userCreate')
-// userCreateForm = addEventListener('submit')
