--- views/index.ejs
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>My Favorite Color</title>
-</head>
-<script>
-    const changeColor = ()=> { 
-        // get select element
-        // get its value 
-        // insert into next line
-        let colorSelect = document.getElementById('colorSelect');
-        let color = colorSelect.value;
-        document.getElementById('body').style.backgroundColor = color;
-        }
-</script>
-<body id="body">
-    <h1>My Favorite Color</h1>
-    <!-- TO DO: -->
-    <!-- add in readout of usename if a user has been selected -->
-    <!-- if a user has not been selected, add in default select for color otherwise -->
-    <!-- if a user has been selected, add these values as default into the creation bar -->
-    <form action="/" id="userCreate" method="POST">
-        <input type="text" name="username" placeholder="username">
-        <select name="color" id="colorSelect">
-            <% for(let color of colors) {%> 
-                <option value="<%=color%>"><%=color%></option>
-            <% } %> 
-        </select>
-        <button onclick="changeColor()">Submit</button>
-        
-    </form>
-    <!-- <button onclick="changeColor()">Submit</button> -->
-    <!-- <script src="userCreate.js"></script> -->
-</body>
-</html>
